refactor dto model mappers
