+++ PReek/Services/DtoModelMapper/toRepository.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+func toRepository(repository: PullRequestDto.Repository) -> Repository {
+    Repository(
+        name: repository.nameWithOwner,
+        url: URL(string: repository.url) ?? URL(string: "https://invalid.data")!
+    )
+}
